+++ booking/src/main/java/com/github/wikicode96/booking/service/BookingService.java
@@ -8,7 +8,7 @@ public interface BookingService {
     BookingEntity newBooking(BookingEntity booking);
     BookingEntity getBookingById(int id);
     List<BookingEntity> getAllBookings();
-    List<BookingEntity> getAllBookingsByUserId(int idUser);
+    List<BookingEntity> getAllBookingsByUserId(Long idUser);
     BookingEntity updateBooking(BookingEntity booking);
     BookingEntity deleteBooking(BookingEntity booking);
 }
